+++ java/src/name/fraser/neil/plaintext/diff_match_patch.java
@@ -81,12 +81,12 @@ public class diff_match_patch {
    * Internal class for returning results from diff_linesToChars().
    * Other less paranoid languages just use a three-element array.
    */
-  protected static class LinesToCharsResult {
-    protected String chars1;
-    protected String chars2;
-    protected List<String> lineArray;
+  public static class LinesToCharsResult {
+    public String chars1;
+    public String chars2;
+    public List<String> lineArray;
 
-    protected LinesToCharsResult(String chars1, String chars2,
+    public LinesToCharsResult(String chars1, String chars2,
         List<String> lineArray) {
       this.chars1 = chars1;
       this.chars2 = chars2;
@@ -500,7 +500,7 @@ public class diff_match_patch {
    *     the List of unique strings.  The zeroth element of the List of
    *     unique strings is intentionally blank.
    */
-  protected LinesToCharsResult diff_linesToChars(String text1, String text2) {
+  public LinesToCharsResult diff_linesToChars(String text1, String text2) {
     List<String> lineArray = new ArrayList<String>();
     Map<String, Integer> lineHash = new HashMap<String, Integer>();
     // e.g. linearray[4] == "Hello\n"
@@ -565,7 +565,7 @@ public class diff_match_patch {
    * @param diffs List of Diff objects.
    * @param lineArray List of unique strings.
    */
-  protected void diff_charsToLines(List<Diff> diffs,
+  public void diff_charsToLines(List<Diff> diffs,
                                   List<String> lineArray) {
     StringBuilder text;
     for (Diff diff : diffs) {
